requirements.txt
uvicorn
fastapi
python-dotenv
pydantic_settings
openai
asyncio